+++ src/GivingAssistant.UserMatchCalculator/FunctionHandler.cs
@@ -71,7 +71,7 @@ namespace GivingAssistant.UserMatchCalculator
                     await createUserTagMatchCommandHandler.Handle(new CreateUserTagMatchCommand
                         {
                             User = user,
-                            Answer = document["SCORE"].AsInt(),
+                            Answer = document["SCORE"].AsDecimal(),
                             Question = questionTagListModel
                         }, CancellationToken.None)
                         ;
